Add files via upload
Added functions load_blocklist and is_blocked.
Updates to thread so that if url is on blocklist will log the request and return HTML page.
Modified format_log_entry's function parameters to include int is_malicious. 1 for yes, 0 for no. 
If it is malicious a tag is added to logstring and then logged.

Still To Do: clean up and more testing.